fix: allow sex to be empty initially
