+++ frontend/src/lib/dogs/validateDogForm.ts
@@ -1,23 +1,4 @@
-import { z } from 'zod';
-
-export const DogFormSchema = z.object({
-  name: z.string().trim().min(1, "Please enter your dog's name"),
-  breed: z.string().trim().min(1, 'Please enter a breed'),
-  dateOfBirth: z.string().trim().min(1, 'Please enter a date of birth'),
-  sex: z
-    .enum(['Male', 'Female'])
-    .or(z.literal('')) // allow empty string in the form
-    .superRefine((v, ctx) => {
-      if (v === '') {
-        ctx.addIssue({
-          code: z.ZodIssueCode.custom,
-          message: 'Please select a sex',
-        });
-      }
-    }),
-});
-
-export type DogFormValues = z.infer<typeof DogFormSchema>;
+import { DogFormSchema, DogFormValues } from './DogFormSchema';
 
 export function validateDogForm(values: DogFormValues): Record<string, string> {
   const result = DogFormSchema.safeParse(values);
