--- frontend/src/types/dog.ts
@@ -1,6 +1,0 @@
-export interface DogFormValues {
-  name: string;
-  breed: string;
-  dateOfBirth: string;
-  sex: string;
-}
